# Conflicts:
#	include/main.h
#	src/Motors/Motor.cpp
#	src/Sensors/Microphone.cpp
#	src/main.cpp

+++ include/Microphone.h
@@ -1,6 +1,7 @@
 /*
- * Microphone.h - Library managing an Ultrasonic Sensor.
+ * Microphone.h - Library for Microphone MAX4466.
  */
+
 #ifndef Microphone_h
 #define Microphone_h
 
@@ -11,13 +12,11 @@ class Microphone
 public:
   Microphone(String id, int AMP_PIN);
   String getJsonSerializedReadings();
-  unsigned long measureAnalog();
+  int measureAnalog();
 
 private:
   String _id;
   int _AMP_PIN;
-  //int _sample_window = 50;
-//   unsigned int _latest_value = 0;
 };
 
 #endif
